requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-annotate --no-hashes --output-file requirements.txt
-e .
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
autoddg @ git+https://github.com/VIDA-NYU/AutoDDG@d70544301a6eba9463a759c7071c463efcba8e8e
beartype==0.22.6
bert-score==0.3.13
certifi==2025.11.12
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6 ; sys_platform == 'win32'
coloredlogs==15.0.1
contourpy==1.3.2 ; python_full_version < '3.11'
contourpy==1.3.3 ; python_full_version >= '3.11'
cycler==0.12.1
datamart-geo==0.3.1
datamart-profiler==0.11
distlib==0.4.0
distro==1.9.0
exceptiongroup==1.3.0 ; python_full_version < '3.11'
fastembed==0.5.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib-metadata==8.7.0
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
kiwisolver==1.4.9
loguru==0.7.3
markupsafe==3.0.3
matplotlib==3.10.7
mmh3==4.1.0
mpmath==1.3.0
networkx==3.4.2 ; python_full_version < '3.11'
networkx==3.6 ; python_full_version >= '3.11'
nltk==3.9.1
nodeenv==1.9.1
numpy==2.2.6 ; python_full_version < '3.11'
numpy==2.3.5 ; python_full_version >= '3.11'
nvidia-cublas-cu12==12.8.4.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cuda-cupti-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cuda-nvrtc-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cuda-runtime-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cudnn-cu12==9.10.2.21 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cufft-cu12==11.3.3.83 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cufile-cu12==1.13.1.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-curand-cu12==10.3.9.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cusolver-cu12==11.7.3.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cusparse-cu12==12.5.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cusparselt-cu12==0.7.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-nccl-cu12==2.27.5 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-nvjitlink-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-nvshmem-cu12==3.3.20 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-nvtx-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
onnxruntime==1.23.2
openai==1.61.0
pypdf==5.1.0
opentelemetry-api==1.38.0
packaging==25.0
pandas==2.3.3
pillow==10.4.0
platformdirs==4.5.0
pre-commit==4.4.0
prometheus-client==0.23.1
protobuf==6.33.1
py-rust-stemmers==0.1.5
pydantic==2.12.4
pydantic-core==2.41.5
pyparsing==3.2.5
pyreadline3==3.5.4 ; sys_platform == 'win32'
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
rank-bm25==0.2.2
regex==2025.11.3
requests==2.32.5
rouge-score==0.1.2
ruff==0.13.0
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.15.3 ; python_full_version < '3.11'
scipy==1.16.3 ; python_full_version >= '3.11'
setuptools==80.9.0 ; python_full_version >= '3.12'
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.35.4
win32-setctime==1.2.0 ; sys_platform == 'win32'
zipp==3.23.0
